refactoring(release): Split the big release.py in multiple components
