+++ tasks/libs/ciproviders/github_api.py
@@ -10,6 +10,9 @@ from functools import lru_cache
 
 import requests
 
+from tasks.libs.common.color import color_message
+from tasks.libs.common.constants import GITHUB_REPO_NAME
+
 try:
     from github import Auth, Github, GithubException, GithubIntegration, GithubObject
     from github.NamedUser import NamedUser
@@ -358,3 +361,67 @@ def get_user_query(login):
     query = '{"query": "query GetUserTeam($login: String!, $org: String!) { user(login: $login) {organization(login: $org) { teams(first:10, userLogins: [$login]){ nodes { slug } } } } }", '
     string_var = f'"variables": {json.dumps(variables)}'
     return query + string_var
+
+
+def create_pr(title, base_branch, target_branch, version, changelog_pr=False):
+    print(color_message("Creating PR", "bold"))
+
+    github = GithubAPI(repository=GITHUB_REPO_NAME)
+
+    # Find milestone based on what the next final version is. If the milestone does not exist, fail.
+    milestone_name = str(version)
+
+    milestone = github.get_milestone_by_name(milestone_name)
+
+    if not milestone or not milestone.number:
+        raise Exit(
+            color_message(
+                f"""Could not find milestone {milestone_name} in the Github repository. Response: {milestone}
+Make sure that milestone is open before trying again.""",
+                "red",
+            ),
+            code=1,
+        )
+
+    pr = github.create_pr(
+        pr_title=title,
+        pr_body="",
+        base_branch=base_branch,
+        target_branch=target_branch,
+    )
+
+    if not pr:
+        raise Exit(
+            color_message(f"Could not create PR in the Github repository. Response: {pr}", "red"),
+            code=1,
+        )
+
+    print(color_message(f"Created PR #{pr.number}", "bold"))
+
+    labels = [
+        "changelog/no-changelog",
+        "qa/no-code-change",
+        "team/agent-build-and-releases",
+        "team/agent-release-management",
+        "category/release_operations",
+    ]
+
+    if changelog_pr:
+        labels.append("backport/main")
+
+    updated_pr = github.update_pr(
+        pull_number=pr.number,
+        milestone_number=milestone.number,
+        labels=labels,
+    )
+
+    if not updated_pr or not updated_pr.number or not updated_pr.html_url:
+        raise Exit(
+            color_message(f"Could not update PR in the Github repository. Response: {updated_pr}", "red"),
+            code=1,
+        )
+
+    print(color_message(f"Set labels and milestone for PR #{updated_pr.number}", "bold"))
+    print(color_message(f"Done creating new PR. Link: {updated_pr.html_url}", "bold"))
+
+    return updated_pr.html_url
