+++ tasks/linter.py
@@ -22,8 +22,10 @@ from tasks.libs.ciproviders.gitlab_api import (
     read_includes,
 )
 from tasks.libs.common.check_tools_version import check_tools_version
+from tasks.libs.common.color import color_message
+from tasks.libs.common.constants import DEFAULT_BRANCH, GITHUB_REPO_NAME
 from tasks.libs.common.git import get_staged_files
-from tasks.libs.common.utils import DEFAULT_BRANCH, GITHUB_REPO_NAME, color_message, is_pr_context, running_in_ci
+from tasks.libs.common.utils import is_pr_context, running_in_ci
 from tasks.libs.types.copyright import CopyrightLinter, LintFailure
 from tasks.modules import GoModule
 from tasks.test_core import ModuleLintResult, process_input_args, process_module_results, test_core
