+++ tasks/omnibus.py
@@ -13,7 +13,8 @@ from tasks.libs.common.omnibus import (
     send_cache_miss_event,
     should_retry_bundle_install,
 )
-from tasks.libs.common.utils import collapsed_section, get_version, load_release_versions, timed
+from tasks.libs.common.utils import collapsed_section, timed
+from tasks.libs.releasing.version import get_version, load_release_versions
 from tasks.ssm import get_pfx_pass, get_signing_cert
