+++ tasks/pipeline.py
@@ -14,15 +14,11 @@ from invoke.exceptions import Exit
 from tasks.libs.ciproviders.github_api import GithubAPI
 from tasks.libs.ciproviders.gitlab_api import get_gitlab_bot_token, get_gitlab_repo, refresh_pipeline
 from tasks.libs.common.color import color_message
-from tasks.libs.common.git import get_commit_sha, get_current_branch
+from tasks.libs.common.constants import DEFAULT_BRANCH, GITHUB_REPO_NAME
+from tasks.libs.common.git import check_clean_branch_state, get_commit_sha, get_current_branch
 from tasks.libs.common.utils import (
-    DEFAULT_BRANCH,
-    GITHUB_REPO_NAME,
-    check_clean_branch_state,
     get_all_allowed_repo_branches,
     is_allowed_repo_branch,
-    nightly_entry_for,
-    release_entry_for,
 )
 from tasks.libs.owners.parsing import read_owners
 from tasks.libs.pipeline.notifications import send_slack_message
@@ -34,6 +30,7 @@ from tasks.libs.pipeline.tools import (
     trigger_agent_pipeline,
     wait_for_pipeline,
 )
+from tasks.libs.releasing.documentation import nightly_entry_for, release_entry_for
 
 
 class GitlabReference(yaml.YAMLObject):
