+++ tasks/release.py
@@ -2,51 +2,49 @@
 Release helper tasks
 """
 
-import json
 import os
 import re
 import sys
 import tempfile
-from collections import OrderedDict
 from datetime import date
 from time import sleep
 
 from gitlab import GitlabError
-from invoke import Failure, task
+from invoke import task
 from invoke.exceptions import Exit
 
-from tasks.libs.ciproviders.github_api import GithubAPI
+from tasks.libs.ciproviders.github_api import GithubAPI, create_pr
 from tasks.libs.ciproviders.gitlab_api import get_gitlab_repo
 from tasks.libs.common.color import color_message
-from tasks.libs.common.git import get_current_branch
-from tasks.libs.common.user_interactions import yes_no_question
-from tasks.libs.common.utils import (
+from tasks.libs.common.constants import (
     DEFAULT_BRANCH,
     GITHUB_REPO_NAME,
-    check_clean_branch_state,
-    get_version,
-    nightly_entry_for,
-    release_entry_for,
 )
+from tasks.libs.common.git import check_base_branch, check_clean_branch_state, get_current_branch, try_git_command
+from tasks.libs.common.user_interactions import yes_no_question
 from tasks.libs.pipeline.notifications import DEFAULT_SLACK_CHANNEL, load_and_validate
 from tasks.libs.releasing.documentation import create_release_page, get_release_page_info
-from tasks.libs.types.version import Version
+from tasks.libs.releasing.json import (
+    _get_release_json_value,
+    _load_release_json,
+    _save_release_json,
+    update_release_json,
+)
+from tasks.libs.releasing.notes import _add_dca_prelude, _add_prelude
+from tasks.libs.releasing.version import (
+    MINOR_RC_VERSION_RE,
+    RC_VERSION_RE,
+    VERSION_RE,
+    _create_version_from_match,
+    check_version,
+    current_version,
+    next_final_version,
+    next_rc_version,
+    parse_major_versions,
+)
 from tasks.modules import DEFAULT_MODULES
 from tasks.pipeline import edit_schedule, run
 from tasks.release_metrics.metrics import get_prs_metrics, get_release_lead_time
-
-# Generic version regex. Aims to match:
-# - X.Y.Z
-# - X.Y.Z-rc.t
-# - X.Y.Z-devel
-# - vX.Y(.Z) (security-agent-policies repo)
-VERSION_RE = re.compile(r'(v)?(\d+)[.](\d+)([.](\d+))?(-devel)?(-rc\.(\d+))?')
-
-# Regex matching rc version tag format like 7.50.0-rc.1
-RC_VERSION_RE = re.compile(r'\d+[.]\d+[.]\d+-rc\.\d+')
-
-# Regex matching minor release rc version tag like x.y.0-rc.1 (semver PATCH == 0), but not x.y.1-rc.1 (semver PATCH > 0)
-MINOR_RC_VERSION_RE = re.compile(r'\d+[.]\d+[.]0-rc\.\d+')
 
 UNFREEZE_REPO_AGENT = "datadog-agent"
 UNFREEZE_REPOS = ["omnibus-software", "omnibus-ruby", "datadog-agent-macos-build", UNFREEZE_REPO_AGENT]
@@ -61,299 +59,6 @@ GITLAB_FILES_TO_UPDATE = [
     ".gitlab-ci.yml",
     ".gitlab/notify/notify.yml",
 ]
-
-
-@task
-def add_prelude(ctx, version):
-    res = ctx.run(f"reno new prelude-release-{version}")
-    new_releasenote = res.stdout.split(' ')[-1].strip()  # get the new releasenote file path
-
-    with open(new_releasenote, "w") as f:
-        f.write(
-            f"""prelude:
-    |
-    Release on: {date.today()}
-
-    - Please refer to the `{version} tag on integrations-core <https://github.com/DataDog/integrations-core/blob/master/AGENT_CHANGELOG.md#datadog-agent-version-{version.replace('.', '')}>`_ for the list of changes on the Core Checks
-"""
-        )
-
-    ctx.run(f"git add {new_releasenote}")
-    print("\nIf not run as part of finish task, commit this with:")
-    print(f"git commit -m \"Add prelude for {version} release\"")
-
-
-@task
-def add_dca_prelude(ctx, agent7_version, agent6_version=""):
-    """
-    Release of the Cluster Agent should be pinned to a version of the Agent.
-    """
-    res = ctx.run(f"reno --rel-notes-dir releasenotes-dca new prelude-release-{agent7_version}")
-    new_releasenote = res.stdout.split(' ')[-1].strip()  # get the new releasenote file path
-
-    if agent6_version != "":
-        agent6_version = (
-            f"--{agent6_version.replace('.', '')}"  # generate the right hyperlink to the agent's changelog.
-        )
-
-    with open(new_releasenote, "w") as f:
-        f.write(
-            f"""prelude:
-    |
-    Released on: {date.today()}
-    Pinned to datadog-agent v{agent7_version}: `CHANGELOG <https://github.com/{GITHUB_REPO_NAME}/blob/{DEFAULT_BRANCH}/CHANGELOG.rst#{agent7_version.replace('.', '')}{agent6_version}>`_."""
-        )
-
-    ctx.run(f"git add {new_releasenote}")
-    print("\nIf not run as part of finish task, commit this with:")
-    print(f"git commit -m \"Add prelude for {agent7_version} release\"")
-
-
-@task
-def add_installscript_prelude(ctx, version):
-    res = ctx.run(f"reno --rel-notes-dir releasenotes-installscript new prelude-release-{version}")
-    new_releasenote = res.stdout.split(' ')[-1].strip()  # get the new releasenote file path
-
-    with open(new_releasenote, "w") as f:
-        f.write(
-            f"""prelude:
-    |
-    Released on: {date.today()}"""
-        )
-
-    ctx.run(f"git add {new_releasenote}")
-    print("\nCommit this with:")
-    print(f"git commit -m \"Add prelude for {version} release\"")
-
-
-@task
-def update_changelog(ctx, new_version=None, target="all", upstream="origin"):
-    """
-    Quick task to generate the new CHANGELOG using reno when releasing a minor
-    version (linux/macOS only).
-    By default generates Agent and Cluster Agent changelogs.
-    Use target == "agent" or target == "cluster-agent" to only generate one or the other.
-    If new_version is omitted, a changelog since last tag on the current branch
-    will be generated.
-    """
-
-    # Step 1 - generate the changelogs
-
-    generate_agent = target in ["all", "agent"]
-    generate_cluster_agent = target in ["all", "cluster-agent"]
-
-    if new_version is not None:
-        new_version_int = list(map(int, new_version.split(".")))
-        if len(new_version_int) != 3:
-            print(f"Error: invalid version: {new_version_int}")
-            raise Exit(1)
-
-        # let's avoid losing uncommitted change with 'git reset --hard'
-        try:
-            ctx.run("git diff --exit-code HEAD", hide="both")
-        except Failure:
-            print("Error: You have uncommitted change, please commit or stash before using update_changelog")
-            return
-
-        # make sure we are up to date
-        ctx.run("git fetch")
-
-        # let's check that the tag for the new version is present
-        try:
-            ctx.run(f"git tag --list | grep {new_version}")
-        except Failure:
-            print(f"Missing '{new_version}' git tag: mandatory to use 'reno'")
-            raise
-
-    if generate_agent:
-        update_changelog_generic(ctx, new_version, "releasenotes", "CHANGELOG.rst")
-    if generate_cluster_agent:
-        update_changelog_generic(ctx, new_version, "releasenotes-dca", "CHANGELOG-DCA.rst")
-
-    # Step 2 - commit changes
-
-    update_branch = f"changelog-update-{new_version}"
-    base_branch = get_current_branch(ctx)
-
-    print(color_message(f"Branching out to {update_branch}", "bold"))
-    ctx.run(f"git checkout -b {update_branch}")
-
-    print(color_message("Committing CHANGELOG.rst and CHANGELOG-DCA.rst", "bold"))
-    print(
-        color_message(
-            "If commit signing is enabled, you will have to make sure the commit gets properly signed.", "bold"
-        )
-    )
-    ctx.run("git add CHANGELOG.rst CHANGELOG-DCA.rst")
-
-    commit_message = f"'Changelog updates for {new_version} release'"
-
-    ok = try_git_command(ctx, f"git commit -m {commit_message}")
-    if not ok:
-        raise Exit(
-            color_message(
-                f"Could not create commit. Please commit manually with:\ngit commit -m {commit_message}\n, push the {update_branch} branch and then open a PR.",
-                "red",
-            ),
-            code=1,
-        )
-
-    # Step 3 - Push and create PR
-
-    print(color_message("Pushing new branch to the upstream repository", "bold"))
-    res = ctx.run(f"git push --set-upstream {upstream} {update_branch}", warn=True)
-    if res.exited is None or res.exited > 0:
-        raise Exit(
-            color_message(
-                f"Could not push branch {update_branch} to the upstream '{upstream}'. Please push it manually and then open a PR.",
-                "red",
-            ),
-            code=1,
-        )
-
-    create_pr(f"Changelog update for {new_version} release", base_branch, update_branch, new_version, changelog_pr=True)
-
-
-def update_changelog_generic(ctx, new_version, changelog_dir, changelog_file):
-    if new_version is None:
-        latest_version = current_version(ctx, 7)
-        ctx.run(f"reno -q --rel-notes-dir {changelog_dir} report --ignore-cache --earliest-version {latest_version}")
-        return
-    new_version_int = list(map(int, new_version.split(".")))
-
-    # removing releasenotes from bugfix on the old minor.
-    branching_point = f"{new_version_int[0]}.{new_version_int[1]}.0-devel"
-    previous_minor = f"{new_version_int[0]}.{new_version_int[1] - 1}"
-    if previous_minor == "7.15":
-        previous_minor = "6.15"  # 7.15 is the first release in the 7.x series
-    log_result = ctx.run(
-        f"git log {branching_point}...remotes/origin/{previous_minor}.x --name-only --oneline | grep {changelog_dir}/notes/ || true"
-    )
-    log_result = log_result.stdout.replace('\n', ' ').strip()
-    if len(log_result) > 0:
-        ctx.run(f"git rm --ignore-unmatch {log_result}")
-
-    # generate the new changelog
-    ctx.run(
-        f"reno --rel-notes-dir {changelog_dir} report --ignore-cache --earliest-version {branching_point} --version {new_version} --no-show-source > /tmp/new_changelog.rst"
-    )
-
-    ctx.run(f"git checkout HEAD -- {changelog_dir}")
-
-    # mac's `sed` has a different syntax for the "-i" paramter
-    # GNU sed has a `--version` parameter while BSD sed does not, using that to do proper detection.
-    try:
-        ctx.run("sed --version", hide='both')
-        sed_i_arg = "-i"
-    except Failure:
-        sed_i_arg = "-i ''"
-    # check whether there is a v6 tag on the same v7 tag, if so add the v6 tag to the release title
-    v6_tag = ""
-    if new_version_int[0] == 7:
-        v6_tag = _find_v6_tag(ctx, new_version)
-        if v6_tag:
-            ctx.run(f"sed {sed_i_arg} -E 's#^{new_version}#{new_version} / {v6_tag}#' /tmp/new_changelog.rst")
-            ctx.run(f"sed {sed_i_arg} -E 's#^======$#================#' /tmp/new_changelog.rst")
-    # remove the old header from the existing changelog
-    ctx.run(f"sed {sed_i_arg} -e '1,4d' {changelog_file}")
-
-    # merging to <changelog_file>
-    ctx.run(f"cat {changelog_file} >> /tmp/new_changelog.rst && mv /tmp/new_changelog.rst {changelog_file}")
-
-    # commit new CHANGELOG
-    ctx.run(f"git add {changelog_file}")
-
-    print("\nCommit this with:")
-    print(f"git commit -m \"Update {changelog_file} for {new_version}\"")
-
-
-@task
-def update_installscript_changelog(ctx, new_version):
-    """
-    Quick task to generate the new CHANGELOG-INSTALLSCRIPT using reno when releasing a minor
-    version (linux/macOS only).
-    """
-    new_version_int = list(map(int, new_version.split(".")))
-
-    if len(new_version_int) != 3:
-        print(f"Error: invalid version: {new_version_int}")
-        raise Exit(1)
-
-    # let's avoid losing uncommitted change with 'git reset --hard'
-    try:
-        ctx.run("git diff --exit-code HEAD", hide="both")
-    except Failure:
-        print("Error: You have uncommitted changes, please commit or stash before using update-installscript-changelog")
-        return
-
-    # make sure we are up to date
-    ctx.run("git fetch")
-
-    # let's check that the tag for the new version is present (needed by reno)
-    try:
-        ctx.run(f"git tag --list | grep installscript-{new_version}")
-    except Failure:
-        print(f"Missing 'installscript-{new_version}' git tag: mandatory to use 'reno'")
-        raise
-
-    # generate the new changelog
-    ctx.run(
-        f"reno --rel-notes-dir releasenotes-installscript report             --ignore-cache             --version installscript-{new_version}             --no-show-source > /tmp/new_changelog-installscript.rst"
-    )
-
-    # reseting git
-    ctx.run("git reset --hard HEAD")
-
-    # mac's `sed` has a different syntax for the "-i" paramter
-    sed_i_arg = "-i"
-    if sys.platform == 'darwin':
-        sed_i_arg = "-i ''"
-    # remove the old header from the existing changelog
-    ctx.run(f"sed {sed_i_arg} -e '1,4d' CHANGELOG-INSTALLSCRIPT.rst")
-
-    if sys.platform != 'darwin':
-        # sed on darwin doesn't support `-z`. On mac, you will need to manually update the following.
-        ctx.run(
-            "sed -z {0} -e 's/installscript-{1}\\n===={2}/{1}\\n{2}/' /tmp/new_changelog-installscript.rst".format(  # noqa: FS002
-                sed_i_arg, new_version, '=' * len(new_version)
-            )
-        )
-
-    # merging to CHANGELOG-INSTALLSCRIPT.rst
-    ctx.run(
-        "cat CHANGELOG-INSTALLSCRIPT.rst >> /tmp/new_changelog-installscript.rst && mv /tmp/new_changelog-installscript.rst CHANGELOG-INSTALLSCRIPT.rst"
-    )
-
-    # commit new CHANGELOG-INSTALLSCRIPT
-    ctx.run("git add CHANGELOG-INSTALLSCRIPT.rst")
-
-    print("\nCommit this with:")
-    print(f"git commit -m \"[INSTALLSCRIPT] Update CHANGELOG-INSTALLSCRIPT for {new_version}\"")
-
-
-@task
-def _find_v6_tag(ctx, v7_tag):
-    """
-    Returns one of the v6 tags that point at the same commit as the passed v7 tag.
-    If none are found, returns the empty string.
-    """
-    v6_tag = ""
-
-    print(f"Looking for a v6 tag pointing to same commit as tag '{v7_tag}'...")
-    # Find commit at which the v7_tag points
-    commit = ctx.run(f"git rev-list --max-count=1 {v7_tag}", hide='out').stdout.strip()
-    try:
-        v6_tags = ctx.run(f"git tag --points-at {commit} | grep -E '^6\\.'", hide='out').stdout.strip().split("\n")
-    except Failure:
-        print(f"Found no v6 tag pointing at same commit as '{v7_tag}'.")
-    else:
-        v6_tag = v6_tags[0]
-        if len(v6_tags) > 1:
-            print(f"Found v6 tags '{v6_tags}', picking {v6_tag}'")
-        else:
-            print(f"Found v6 tag '{v6_tag}'")
-
-    return v6_tag
 
 
 @task
@@ -372,476 +77,6 @@ def list_major_change(_, milestone):
 
     for pr in pull_requests:
         print(f"#{pr.number}: {pr.title} ({pr.html_url})")
-
-
-#
-# release.json manipulation invoke tasks section
-#
-
-##
-## I/O functions
-##
-
-
-def _load_release_json():
-    with open("release.json") as release_json_stream:
-        return json.load(release_json_stream, object_pairs_hook=OrderedDict)
-
-
-def _save_release_json(release_json):
-    with open("release.json", "w") as release_json_stream:
-        # Note, no space after the comma
-        json.dump(release_json, release_json_stream, indent=4, sort_keys=False, separators=(',', ': '))
-        release_json_stream.write('\n')
-
-
-##
-## Utils
-##
-
-
-def _create_version_from_match(match):
-    groups = match.groups()
-    version = Version(
-        major=int(groups[1]),
-        minor=int(groups[2]),
-        patch=int(groups[4]) if groups[4] and groups[4] != 0 else None,
-        devel=True if groups[5] else False,
-        rc=int(groups[7]) if groups[7] and groups[7] != 0 else None,
-        prefix=groups[0] if groups[0] else "",
-    )
-    return version
-
-
-def _stringify_config(config_dict):
-    """
-    Takes a config dict of the following form:
-    {
-        "xxx_VERSION": Version(major: x, minor: y, patch: z, rc: t, prefix: "pre"),
-        "xxx_HASH": "hashvalue",
-        ...
-    }
-
-    and transforms all VERSIONs into their string representation (using the Version object's __str__).
-    """
-    return {key: str(value) for key, value in config_dict.items()}
-
-
-def build_compatible_version_re(allowed_major_versions, minor_version):
-    """
-    Returns a regex that matches only versions whose major version is
-    in the provided list of allowed_major_versions, and whose minor version matches
-    the provided minor version.
-    """
-    return re.compile(
-        r'(v)?({})[.]({})([.](\d+))+(-devel)?(-rc\.(\d+))?(?!-\w)'.format(  # noqa: FS002
-            "|".join(allowed_major_versions), minor_version
-        )
-    )
-
-
-##
-## Base functions to fetch candidate versions on other repositories
-##
-
-
-def _get_highest_repo_version(
-    repo, version_prefix, version_re, allowed_major_versions=None, max_version: Version = None
-):
-    # If allowed_major_versions is not specified, search for all versions by using an empty
-    # major version prefix.
-    if not allowed_major_versions:
-        allowed_major_versions = [""]
-
-    highest_version = None
-
-    gh = GithubAPI(repository=f'Datadog/{repo}')
-
-    for major_version in allowed_major_versions:
-        tags = gh.get_tags(f'{version_prefix}{major_version}')
-
-        for tag in tags:
-            match = version_re.search(tag.name)
-            if match:
-                this_version = _create_version_from_match(match)
-                if max_version:
-                    # Get the max version that corresponds to the major version
-                    # of the current tag
-                    this_max_version = max_version.clone()
-                    this_max_version.major = this_version.major
-                    if this_version > this_max_version:
-                        continue
-                if this_version > highest_version:
-                    highest_version = this_version
-
-        # The allowed_major_versions are listed in order of preference
-        # If something matching a given major version exists, no need to
-        # go through the next ones.
-        if highest_version:
-            break
-
-    if not highest_version:
-        raise Exit(f"Couldn't find any matching {repo} version.", 1)
-
-    return highest_version
-
-
-def _get_release_version_from_release_json(release_json, major_version, version_re, release_json_key=None):
-    """
-    If release_json_key is None, returns the highest version entry in release.json.
-    If release_json_key is set, returns the entry for release_json_key of the highest version entry in release.json.
-    """
-
-    release_version = None
-    release_component_version = None
-
-    # Get the release entry for the given Agent major version
-    release_entry_name = release_entry_for(major_version)
-    release_json_entry = release_json.get(release_entry_name, None)
-
-    # Check that the release entry exists, otherwise fail
-    if release_json_entry:
-        release_version = release_entry_name
-
-        # Check that the component's version is defined in the release entry
-        if release_json_key is not None:
-            match = version_re.match(release_json_entry.get(release_json_key, ""))
-            if match:
-                release_component_version = _create_version_from_match(match)
-            else:
-                print(
-                    f"{release_entry_name} does not have a valid {release_json_key} ({release_json_entry.get(release_json_key, '')}), ignoring"
-                )
-
-    if not release_version:
-        raise Exit(f"Couldn't find any matching {release_version} version.", 1)
-
-    if release_json_key is not None:
-        return release_component_version
-
-    return release_version
-
-
-##
-## Variables needed for the repository version fetch functions
-##
-
-# COMPATIBLE_MAJOR_VERSIONS lists the major versions of tags
-# that can be used with a given Agent version
-# This is here for compatibility and simplicity reasons, as in most repos
-# we don't create both 6 and 7 tags for a combined Agent 6 & 7 release.
-# The order matters, eg. when fetching matching tags for an Agent 6 entry,
-# tags starting with 6 will be preferred to tags starting with 7.
-COMPATIBLE_MAJOR_VERSIONS = {6: ["6", "7"], 7: ["7"]}
-
-
-# Message templates for the below functions
-# Defined here either because they're long and would make the code less legible,
-# or because they're used multiple times.
-RC_TAG_QUESTION_TEMPLATE = "The {} tag found is an RC tag: {}. Are you sure you want to use it?"
-TAG_FOUND_TEMPLATE = "The {} tag is {}"
-
-
-##
-## Repository version fetch functions
-## The following functions aim at returning the correct version to use for a given
-## repository, after compatibility & user confirmations
-## The version object returned by such functions should be ready to be used to fill
-## the release.json entry.
-##
-
-
-def _fetch_dependency_repo_version(
-    repo_name,
-    new_agent_version,
-    max_agent_version,
-    allowed_major_versions,
-    compatible_version_re,
-    check_for_rc,
-):
-    """
-    Fetches the latest tag on a given repository whose version scheme matches the one used for the Agent,
-    with the following constraints:
-    - the tag must have a major version that's in allowed_major_versions
-    - the tag must match compatible_version_re (the main usage is to restrict the compatible tags to the
-      ones with the same minor version as the Agent)?
-
-    If check_for_rc is true, a warning will be emitted if the latest version that satisfies
-    the constraints is an RC. User confirmation is then needed to check that this is desired.
-    """
-
-    # Get the highest repo version that's not higher than the Agent version we're going to build
-    # We don't want to use a tag on dependent repositories that is supposed to be used in a future
-    # release of the Agent (eg. if 7.31.1-rc.1 is tagged on integrations-core while we're releasing 7.30.0).
-    version = _get_highest_repo_version(
-        repo_name,
-        new_agent_version.prefix,
-        compatible_version_re,
-        allowed_major_versions,
-        max_version=max_agent_version,
-    )
-
-    if check_for_rc and version.is_rc():
-        if not yes_no_question(RC_TAG_QUESTION_TEMPLATE.format(repo_name, version), "orange", False):
-            raise Exit("Aborting release.json update.", 1)
-
-    print(TAG_FOUND_TEMPLATE.format(repo_name, version))
-    return version
-
-
-def _get_jmxfetch_release_json_info(release_json, agent_major_version, is_first_rc=False):
-    """
-    Gets the JMXFetch version info from the previous entries in the release.json file.
-    """
-
-    release_json_version_data = _get_release_json_info_for_next_rc(release_json, agent_major_version, is_first_rc)
-
-    jmxfetch_version = release_json_version_data['JMXFETCH_VERSION']
-    jmxfetch_shasum = release_json_version_data['JMXFETCH_HASH']
-
-    print(f"The JMXFetch version is {jmxfetch_version}")
-
-    return jmxfetch_version, jmxfetch_shasum
-
-
-def _get_windows_release_json_info(release_json, agent_major_version, is_first_rc=False):
-    """
-    Gets the Windows NPM driver info from the previous entries in the release.json file.
-    """
-    release_json_version_data = _get_release_json_info_for_next_rc(release_json, agent_major_version, is_first_rc)
-
-    win_ddnpm_driver, win_ddnpm_version, win_ddnpm_shasum = _get_windows_driver_info(release_json_version_data, 'DDNPM')
-    win_ddprocmon_driver, win_ddprocmon_version, win_ddprocmon_shasum = _get_windows_driver_info(
-        release_json_version_data, 'DDPROCMON'
-    )
-
-    return (
-        win_ddnpm_driver,
-        win_ddnpm_version,
-        win_ddnpm_shasum,
-        win_ddprocmon_driver,
-        win_ddprocmon_version,
-        win_ddprocmon_shasum,
-    )
-
-
-def _get_windows_driver_info(release_json_version_data, driver_name):
-    """
-    Gets the Windows driver info from the release.json version data.
-    """
-    driver_key = f'WINDOWS_{driver_name}_DRIVER'
-    version_key = f'WINDOWS_{driver_name}_VERSION'
-    shasum_key = f'WINDOWS_{driver_name}_SHASUM'
-
-    driver_value = release_json_version_data[driver_key]
-    version_value = release_json_version_data[version_key]
-    shasum_value = release_json_version_data[shasum_key]
-
-    if driver_value not in ['release-signed', 'attestation-signed']:
-        print(f"WARN: {driver_key} value '{driver_value}' is not valid")
-
-    print(f"The windows {driver_name.lower()} version is {version_value}")
-
-    return driver_value, version_value, shasum_value
-
-
-def _get_release_json_info_for_next_rc(release_json, agent_major_version, is_first_rc=False):
-    """
-    Gets the version info from the previous entries in the release.json file.
-    """
-
-    # First RC should use the data from nightly section otherwise reuse the last RC info
-    if is_first_rc:
-        previous_release_json_version = nightly_entry_for(agent_major_version)
-    else:
-        previous_release_json_version = release_entry_for(agent_major_version)
-
-    print(f"Using '{previous_release_json_version}' values")
-
-    return release_json[previous_release_json_version]
-
-
-##
-## release_json object update function
-##
-
-
-def _update_release_json_entry(
-    release_json,
-    release_entry,
-    integrations_version,
-    omnibus_software_version,
-    omnibus_ruby_version,
-    jmxfetch_version,
-    jmxfetch_shasum,
-    security_agent_policies_version,
-    macos_build_version,
-    windows_ddnpm_driver,
-    windows_ddnpm_version,
-    windows_ddnpm_shasum,
-    windows_ddprocmon_driver,
-    windows_ddprocmon_version,
-    windows_ddprocmon_shasum,
-):
-    """
-    Adds a new entry to provided release_json object with the provided parameters, and returns the new release_json object.
-    """
-
-    print(f"Jmxfetch's SHA256 is {jmxfetch_shasum}")
-    print(f"Windows DDNPM's SHA256 is {windows_ddnpm_shasum}")
-    print(f"Windows DDPROCMON's SHA256 is {windows_ddprocmon_shasum}")
-
-    new_version_config = OrderedDict()
-    new_version_config["INTEGRATIONS_CORE_VERSION"] = integrations_version
-    new_version_config["OMNIBUS_SOFTWARE_VERSION"] = omnibus_software_version
-    new_version_config["OMNIBUS_RUBY_VERSION"] = omnibus_ruby_version
-    new_version_config["JMXFETCH_VERSION"] = jmxfetch_version
-    new_version_config["JMXFETCH_HASH"] = jmxfetch_shasum
-    new_version_config["SECURITY_AGENT_POLICIES_VERSION"] = security_agent_policies_version
-    new_version_config["MACOS_BUILD_VERSION"] = macos_build_version
-    new_version_config["WINDOWS_DDNPM_DRIVER"] = windows_ddnpm_driver
-    new_version_config["WINDOWS_DDNPM_VERSION"] = windows_ddnpm_version
-    new_version_config["WINDOWS_DDNPM_SHASUM"] = windows_ddnpm_shasum
-    new_version_config["WINDOWS_DDPROCMON_DRIVER"] = windows_ddprocmon_driver
-    new_version_config["WINDOWS_DDPROCMON_VERSION"] = windows_ddprocmon_version
-    new_version_config["WINDOWS_DDPROCMON_SHASUM"] = windows_ddprocmon_shasum
-
-    # Necessary if we want to maintain the JSON order, so that humans don't get confused
-    new_release_json = OrderedDict()
-
-    # Add all versions from the old release.json
-    for key, value in release_json.items():
-        new_release_json[key] = value
-
-    # Then update the entry
-    new_release_json[release_entry] = _stringify_config(new_version_config)
-
-    return new_release_json
-
-
-##
-## Main functions
-##
-
-
-def _update_release_json(release_json, release_entry, new_version: Version, max_version: Version):
-    """
-    Updates the provided release.json object by fetching compatible versions for all dependencies
-    of the provided Agent version, constructing the new entry, adding it to the release.json object
-    and returning it.
-    """
-
-    allowed_major_versions = COMPATIBLE_MAJOR_VERSIONS[new_version.major]
-
-    # Part 1: repositories which follow the Agent version scheme
-
-    # For repositories which follow the Agent version scheme, we want to only get
-    # tags with the same minor version, to avoid problems when releasing a patch
-    # version while a minor version release is ongoing.
-    compatible_version_re = build_compatible_version_re(allowed_major_versions, new_version.minor)
-
-    # If the new version is a final version, set the check_for_rc flag to true to warn if a dependency's version
-    # is an RC.
-    check_for_rc = not new_version.is_rc()
-
-    integrations_version = _fetch_dependency_repo_version(
-        "integrations-core",
-        new_version,
-        max_version,
-        allowed_major_versions,
-        compatible_version_re,
-        check_for_rc,
-    )
-
-    omnibus_software_version = _fetch_dependency_repo_version(
-        "omnibus-software",
-        new_version,
-        max_version,
-        allowed_major_versions,
-        compatible_version_re,
-        check_for_rc,
-    )
-
-    omnibus_ruby_version = _fetch_dependency_repo_version(
-        "omnibus-ruby",
-        new_version,
-        max_version,
-        allowed_major_versions,
-        compatible_version_re,
-        check_for_rc,
-    )
-
-    macos_build_version = _fetch_dependency_repo_version(
-        "datadog-agent-macos-build",
-        new_version,
-        max_version,
-        allowed_major_versions,
-        compatible_version_re,
-        check_for_rc,
-    )
-
-    # Part 2: repositories which have their own version scheme
-
-    # jmxfetch version is updated directly by the AML team
-    jmxfetch_version, jmxfetch_shasum = _get_jmxfetch_release_json_info(
-        release_json, new_version.major, is_first_rc=(new_version.rc == 1)
-    )
-
-    # security agent policies are updated directly by the CWS team
-    security_agent_policies_version = _get_release_version_from_release_json(
-        release_json, new_version.major, VERSION_RE, "SECURITY_AGENT_POLICIES_VERSION"
-    )
-    print(TAG_FOUND_TEMPLATE.format("security-agent-policies", security_agent_policies_version))
-
-    (
-        windows_ddnpm_driver,
-        windows_ddnpm_version,
-        windows_ddnpm_shasum,
-        windows_ddprocmon_driver,
-        windows_ddprocmon_version,
-        windows_ddprocmon_shasum,
-    ) = _get_windows_release_json_info(release_json, new_version.major, is_first_rc=(new_version.rc == 1))
-
-    # Add new entry to the release.json object and return it
-    return _update_release_json_entry(
-        release_json,
-        release_entry,
-        integrations_version,
-        omnibus_software_version,
-        omnibus_ruby_version,
-        jmxfetch_version,
-        jmxfetch_shasum,
-        security_agent_policies_version,
-        macos_build_version,
-        windows_ddnpm_driver,
-        windows_ddnpm_version,
-        windows_ddnpm_shasum,
-        windows_ddprocmon_driver,
-        windows_ddprocmon_version,
-        windows_ddprocmon_shasum,
-    )
-
-
-def update_release_json(new_version: Version, max_version: Version):
-    """
-    Updates the release entries in release.json to prepare the next RC or final build.
-    """
-    release_json = _load_release_json()
-
-    release_entry = release_entry_for(new_version.major)
-    print(f"Updating {release_entry} for {new_version}")
-
-    # Update release.json object with the entry for the new version
-    release_json = _update_release_json(release_json, release_entry, new_version, max_version)
-
-    _save_release_json(release_json)
-
-
-def check_version(agent_version):
-    """Check Agent version to see if it is valid."""
-    version_re = re.compile(r'7[.](\d+)[.](\d+)(-rc\.(\d+))?')
-    if not version_re.match(agent_version):
-        raise Exit(message="Version should be of the form 7.Y.Z or 7.Y.Z-rc.t")
 
 
 @task
@@ -957,91 +192,6 @@ def tag_devel(ctx, agent_version, commit="HEAD", verify=True, push=True, force=F
     tag_modules(ctx, agent_version, commit, verify, push, force, devel=True)
 
 
-def current_version(ctx, major_version) -> Version:
-    return _create_version_from_match(VERSION_RE.search(get_version(ctx, major_version=major_version, release=True)))
-
-
-def next_final_version(ctx, major_version, patch_version) -> Version:
-    previous_version = current_version(ctx, major_version)
-
-    # Set the new version
-    if previous_version.is_devel():
-        # If the previous version was a devel version, use the same version without devel
-        # (should never happen during regular releases, we always do at least one RC)
-        return previous_version.non_devel_version()
-    if previous_version.is_rc():
-        # If the previous version was an RC version, use the same version without RC
-        return previous_version.next_version(rc=False)
-
-    # Else, the latest version was a final release, so we use the next release
-    # (eg. 7.32.1 from 7.32.0).
-    if patch_version:
-        return previous_version.next_version(bump_patch=True, rc=False)
-    else:
-        return previous_version.next_version(bump_minor=True, rc=False)
-
-
-def next_rc_version(ctx, major_version, patch_version=False) -> Version:
-    # Fetch previous version from the most recent tag on the branch
-    previous_version = current_version(ctx, major_version)
-
-    if previous_version.is_rc():
-        # We're already on an RC, only bump the RC version
-        new_version = previous_version.next_version(rc=True)
-    else:
-        if patch_version:
-            new_version = previous_version.next_version(bump_patch=True, rc=True)
-        else:
-            # Minor version bump, we're doing a standard release:
-            # - if the previous tag is a devel tag, use it without the devel tag
-            # - otherwise (should not happen during regular release cycles), bump the minor version
-            if previous_version.is_devel():
-                new_version = previous_version.non_devel_version()
-                new_version = new_version.next_version(rc=True)
-            else:
-                new_version = previous_version.next_version(bump_minor=True, rc=True)
-
-    return new_version
-
-
-def check_base_branch(branch, release_version):
-    """
-    Checks if the given branch is either the default branch or the release branch associated
-    with the given release version.
-    """
-    return branch == DEFAULT_BRANCH or branch == release_version.branch()
-
-
-def parse_major_versions(major_versions):
-    return sorted(int(x) for x in major_versions.split(","))
-
-
-def try_git_command(ctx, git_command):
-    """
-    Try a git command that should be retried (after user confirmation) if it fails.
-    Primarily useful for commands which can fail if commit signing fails: we don't want the
-    whole workflow to fail if that happens, we want to retry.
-    """
-
-    do_retry = True
-
-    while do_retry:
-        res = ctx.run(git_command, warn=True)
-        if res.exited is None or res.exited > 0:
-            print(
-                color_message(
-                    f"Failed to run \"{git_command}\" (did the commit/tag signing operation fail?)",
-                    "orange",
-                )
-            )
-            do_retry = yes_no_question("Do you want to retry this operation?", color="orange", default=True)
-            continue
-
-        return True
-
-    return False
-
-
 @task
 def finish(ctx, major_versions="6,7", upstream="origin"):
     """
@@ -1105,10 +255,10 @@ def finish(ctx, major_versions="6,7", upstream="origin"):
 
     # Step 4: Add release changelog preludes
     print(color_message("Adding Agent release changelog prelude", "bold"))
-    add_prelude(ctx, str(new_version))
+    _add_prelude(ctx, str(new_version))
 
     print(color_message("Adding DCA release changelog prelude", "bold"))
-    add_dca_prelude(ctx, str(new_version))
+    _add_dca_prelude(ctx, str(new_version))
 
     ok = try_git_command(ctx, f"git commit -m 'Add preludes for {new_version} release'")
     if not ok:
@@ -1276,70 +426,6 @@ def create_rc(ctx, major_versions="6,7", patch_version=False, upstream="origin",
         )
 
 
-def create_pr(title, base_branch, target_branch, version, changelog_pr=False):
-    print(color_message("Creating PR", "bold"))
-
-    github = GithubAPI(repository=GITHUB_REPO_NAME)
-
-    # Find milestone based on what the next final version is. If the milestone does not exist, fail.
-    milestone_name = str(version)
-
-    milestone = github.get_milestone_by_name(milestone_name)
-
-    if not milestone or not milestone.number:
-        raise Exit(
-            color_message(
-                f"""Could not find milestone {milestone_name} in the Github repository. Response: {milestone}
-Make sure that milestone is open before trying again.""",
-                "red",
-            ),
-            code=1,
-        )
-
-    pr = github.create_pr(
-        pr_title=title,
-        pr_body="",
-        base_branch=base_branch,
-        target_branch=target_branch,
-    )
-
-    if not pr:
-        raise Exit(
-            color_message(f"Could not create PR in the Github repository. Response: {pr}", "red"),
-            code=1,
-        )
-
-    print(color_message(f"Created PR #{pr.number}", "bold"))
-
-    labels = [
-        "changelog/no-changelog",
-        "qa/no-code-change",
-        "team/agent-build-and-releases",
-        "team/agent-release-management",
-        "category/release_operations",
-    ]
-
-    if changelog_pr:
-        labels.append("backport/main")
-
-    updated_pr = github.update_pr(
-        pull_number=pr.number,
-        milestone_number=milestone.number,
-        labels=labels,
-    )
-
-    if not updated_pr or not updated_pr.number or not updated_pr.html_url:
-        raise Exit(
-            color_message(f"Could not update PR in the Github repository. Response: {updated_pr}", "red"),
-            code=1,
-        )
-
-    print(color_message(f"Set labels and milestone for PR #{updated_pr.number}", "bold"))
-    print(color_message(f"Done creating new PR. Link: {updated_pr.html_url}", "bold"))
-
-    return updated_pr.html_url
-
-
 @task
 def build_rc(ctx, major_versions="6,7", patch_version=False, k8s_deployments=False):
     """
@@ -1447,20 +533,6 @@ def set_release_json(_, key, value):
 def get_release_json_value(_, key):
     release_json = _get_release_json_value(key)
     print(release_json)
-
-
-def _get_release_json_value(key):
-    release_json = _load_release_json()
-
-    path = key.split('::')
-
-    for element in path:
-        if element not in release_json:
-            raise Exit(code=1, message=f"Couldn't find '{key}' in release.json")
-
-        release_json = release_json.get(element)
-
-    return release_json
 
 
 def create_and_update_release_branch(ctx, repo, release_branch, base_directory="~/dd", upstream="origin"):
