+++ tasks/unit-tests/release_tests.py
@@ -8,6 +8,16 @@ from unittest import mock
 from invoke.exceptions import Exit
 
 from tasks import release
+from tasks.libs.releasing.documentation import nightly_entry_for, release_entry_for
+from tasks.libs.releasing.json import (
+    COMPATIBLE_MAJOR_VERSIONS,
+    _get_jmxfetch_release_json_info,
+    _get_release_json_info_for_next_rc,
+    _get_release_version_from_release_json,
+    _get_windows_release_json_info,
+    _update_release_json_entry,
+)
+from tasks.libs.releasing.version import _get_highest_repo_version, build_compatible_version_re
 from tasks.libs.types.version import Version
 
 
@@ -62,83 +72,83 @@ def mocked_github_requests_incorrect_get(*_args, **_kwargs):
 
 
 class TestGetHighestRepoVersion(unittest.TestCase):
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_ignore_incorrect_tag(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_incorrect_get
         gh_mock.return_value = gh_instance
-        version = release._get_highest_repo_version(
+        version = _get_highest_repo_version(
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[7], 28),
-            release.COMPATIBLE_MAJOR_VERSIONS[7],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[7], 28),
+            COMPATIBLE_MAJOR_VERSIONS[7],
         )
         self.assertEqual(version, Version(major=7, minor=28, patch=0, rc=2))
 
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_one_allowed_major_multiple_entries(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_get
         gh_mock.return_value = gh_instance
-        version = release._get_highest_repo_version(
+        version = _get_highest_repo_version(
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[7], 28),
-            release.COMPATIBLE_MAJOR_VERSIONS[7],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[7], 28),
+            COMPATIBLE_MAJOR_VERSIONS[7],
         )
         self.assertEqual(version, Version(major=7, minor=28, patch=1))
 
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_one_allowed_major_one_entry(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_get
         gh_mock.return_value = gh_instance
-        version = release._get_highest_repo_version(
+        version = _get_highest_repo_version(
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[7], 29),
-            release.COMPATIBLE_MAJOR_VERSIONS[7],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[7], 29),
+            COMPATIBLE_MAJOR_VERSIONS[7],
         )
         self.assertEqual(version, Version(major=7, minor=29, patch=0))
 
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_multiple_allowed_majors_multiple_entries(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_get
         gh_mock.return_value = gh_instance
-        version = release._get_highest_repo_version(
+        version = _get_highest_repo_version(
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[6], 28),
-            release.COMPATIBLE_MAJOR_VERSIONS[6],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[6], 28),
+            COMPATIBLE_MAJOR_VERSIONS[6],
         )
         self.assertEqual(version, Version(major=6, minor=28, patch=1))
 
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_multiple_allowed_majors_one_entry(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_get
         gh_mock.return_value = gh_instance
-        version = release._get_highest_repo_version(
+        version = _get_highest_repo_version(
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[6], 29),
-            release.COMPATIBLE_MAJOR_VERSIONS[6],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[6], 29),
+            COMPATIBLE_MAJOR_VERSIONS[6],
         )
         self.assertEqual(version, Version(major=6, minor=29, patch=0))
 
-    @mock.patch('tasks.release.GithubAPI')
+    @mock.patch('tasks.libs.releasing.version.GithubAPI')
     def test_nonexistant_minor(self, gh_mock):
         gh_instance = mock.MagicMock()
         gh_instance.get_tags.side_effect = mocked_github_requests_get
         gh_mock.return_value = gh_instance
         self.assertRaises(
             Exit,
-            release._get_highest_repo_version,
+            _get_highest_repo_version,
             "target-repo",
             "",
-            release.build_compatible_version_re(release.COMPATIBLE_MAJOR_VERSIONS[7], 30),
-            release.COMPATIBLE_MAJOR_VERSIONS[7],
+            build_compatible_version_re(COMPATIBLE_MAJOR_VERSIONS[7], 30),
+            COMPATIBLE_MAJOR_VERSIONS[7],
         )
 
 
@@ -160,7 +170,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
         self.maxDiff = None
         initial_release_json = OrderedDict(
             {
-                release.nightly_entry_for(6): {
+                nightly_entry_for(6): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -175,7 +185,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.nightly_entry_for(7): {
+                nightly_entry_for(7): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -190,7 +200,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.release_entry_for(6): {
+                release_entry_for(6): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -205,7 +215,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.release_entry_for(7): {
+                release_entry_for(7): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -237,9 +247,9 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
         windows_ddprocmon_version = "1.2.1"
         windows_ddprocmon_shasum = "windowsddprocmonshasum"
 
-        release_json = release._update_release_json_entry(
+        release_json = _update_release_json_entry(
             release_json=initial_release_json,
-            release_entry=release.release_entry_for(7),
+            release_entry=release_entry_for(7),
             integrations_version=integrations_version,
             omnibus_ruby_version=omnibus_ruby_version,
             omnibus_software_version=omnibus_software_version,
@@ -257,7 +267,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
 
         expected_release_json = OrderedDict(
             {
-                release.nightly_entry_for(6): {
+                nightly_entry_for(6): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -272,7 +282,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.nightly_entry_for(7): {
+                nightly_entry_for(7): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -287,7 +297,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.release_entry_for(6): {
+                release_entry_for(6): {
                     "INTEGRATIONS_CORE_VERSION": "master",
                     "OMNIBUS_SOFTWARE_VERSION": "master",
                     "OMNIBUS_RUBY_VERSION": "datadog-5.5.0",
@@ -302,7 +312,7 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
                     "WINDOWS_DDPROCMON_VERSION": "0.98.2.git.86.53d1ee4",
                     "WINDOWS_DDPROCMON_SHASUM": "5d31cbf7aea921edd5ba34baf074e496749265a80468b65a034d3796558a909e",
                 },
-                release.release_entry_for(7): {
+                release_entry_for(7): {
                     "INTEGRATIONS_CORE_VERSION": str(integrations_version),
                     "OMNIBUS_SOFTWARE_VERSION": str(omnibus_software_version),
                     "OMNIBUS_RUBY_VERSION": str(omnibus_ruby_version),
@@ -325,40 +335,40 @@ class TestUpdateReleaseJsonEntry(unittest.TestCase):
 
 class TestGetReleaseVersionFromReleaseJson(unittest.TestCase):
     test_release_json = {
-        release.nightly_entry_for(6): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "master"},
-        release.nightly_entry_for(7): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "master"},
-        release.release_entry_for(6): {"JMXFETCH_VERSION": "0.43.0", "SECURITY_AGENT_POLICIES_VERSION": "v0.10"},
-        release.release_entry_for(7): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "v0.10"},
+        nightly_entry_for(6): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "master"},
+        nightly_entry_for(7): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "master"},
+        release_entry_for(6): {"JMXFETCH_VERSION": "0.43.0", "SECURITY_AGENT_POLICIES_VERSION": "v0.10"},
+        release_entry_for(7): {"JMXFETCH_VERSION": "0.44.1", "SECURITY_AGENT_POLICIES_VERSION": "v0.10"},
     }
 
     def test_release_version_6(self):
-        version = release._get_release_version_from_release_json(self.test_release_json, 6, release.VERSION_RE)
-        self.assertEqual(version, release.release_entry_for(6))
+        version = _get_release_version_from_release_json(self.test_release_json, 6, release.VERSION_RE)
+        self.assertEqual(version, release_entry_for(6))
 
     def test_release_version_7(self):
-        version = release._get_release_version_from_release_json(self.test_release_json, 7, release.VERSION_RE)
-        self.assertEqual(version, release.release_entry_for(7))
+        version = _get_release_version_from_release_json(self.test_release_json, 7, release.VERSION_RE)
+        self.assertEqual(version, release_entry_for(7))
 
     def test_release_jmxfetch_version_6(self):
-        version = release._get_release_version_from_release_json(
+        version = _get_release_version_from_release_json(
             self.test_release_json, 6, release.VERSION_RE, release_json_key="JMXFETCH_VERSION"
         )
         self.assertEqual(version, Version(major=0, minor=43, patch=0))
 
     def test_release_jmxfetch_version_7(self):
-        version = release._get_release_version_from_release_json(
+        version = _get_release_version_from_release_json(
             self.test_release_json, 7, release.VERSION_RE, release_json_key="JMXFETCH_VERSION"
         )
         self.assertEqual(version, Version(major=0, minor=44, patch=1))
 
     def test_release_security_version_6(self):
-        version = release._get_release_version_from_release_json(
+        version = _get_release_version_from_release_json(
             self.test_release_json, 6, release.VERSION_RE, release_json_key="SECURITY_AGENT_POLICIES_VERSION"
         )
         self.assertEqual(version, Version(prefix="v", major=0, minor=10))
 
     def test_release_security_version_7(self):
-        version = release._get_release_version_from_release_json(
+        version = _get_release_version_from_release_json(
             self.test_release_json, 7, release.VERSION_RE, release_json_key="SECURITY_AGENT_POLICIES_VERSION"
         )
         self.assertEqual(version, Version(prefix="v", major=0, minor=10))
@@ -366,7 +376,7 @@ class TestGetReleaseVersionFromReleaseJson(unittest.TestCase):
 
 class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
     test_release_json = {
-        release.nightly_entry_for(6): {
+        nightly_entry_for(6): {
             "WINDOWS_DDNPM_DRIVER": "attestation-signed",
             "WINDOWS_DDNPM_VERSION": "nightly-ddnpm-version",
             "WINDOWS_DDNPM_SHASUM": "nightly-ddnpm-sha",
@@ -374,7 +384,7 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
             "WINDOWS_DDPROCMON_VERSION": "nightly-ddprocmon-version",
             "WINDOWS_DDPROCMON_SHASUM": "nightly-ddprocmon-sha",
         },
-        release.nightly_entry_for(7): {
+        nightly_entry_for(7): {
             "WINDOWS_DDNPM_DRIVER": "attestation-signed",
             "WINDOWS_DDNPM_VERSION": "nightly-ddnpm-version",
             "WINDOWS_DDNPM_SHASUM": "nightly-ddnpm-sha",
@@ -382,7 +392,7 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
             "WINDOWS_DDPROCMON_VERSION": "nightly-ddprocmon-version",
             "WINDOWS_DDPROCMON_SHASUM": "nightly-ddprocmon-sha",
         },
-        release.release_entry_for(6): {
+        release_entry_for(6): {
             "WINDOWS_DDNPM_DRIVER": "release-signed",
             "WINDOWS_DDNPM_VERSION": "rc3-ddnpm-version",
             "WINDOWS_DDNPM_SHASUM": "rc3-ddnpm-sha",
@@ -390,7 +400,7 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
             "WINDOWS_DDPROCMON_VERSION": "rc3-ddprocmon-version",
             "WINDOWS_DDPROCMON_SHASUM": "rc3-ddprocmon-sha",
         },
-        release.release_entry_for(7): {
+        release_entry_for(7): {
             "WINDOWS_DDNPM_DRIVER": "release-signed",
             "WINDOWS_DDNPM_VERSION": "rc3-ddnpm-version",
             "WINDOWS_DDNPM_SHASUM": "rc3-ddnpm-sha",
@@ -408,7 +418,7 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
             ddprocmon_driver,
             ddprocmon_version,
             ddprocmon_shasum,
-        ) = release._get_windows_release_json_info(self.test_release_json, 7, True)
+        ) = _get_windows_release_json_info(self.test_release_json, 7, True)
 
         self.assertEqual(ddnpm_driver, 'attestation-signed')
         self.assertEqual(ddnpm_version, 'nightly-ddnpm-version')
@@ -425,7 +435,7 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
             ddprocmon_driver,
             ddprocmon_version,
             ddprocmon_shasum,
-        ) = release._get_windows_release_json_info(self.test_release_json, 7, False)
+        ) = _get_windows_release_json_info(self.test_release_json, 7, False)
 
         self.assertEqual(ddnpm_driver, 'release-signed')
         self.assertEqual(ddnpm_version, 'rc3-ddnpm-version')
@@ -437,26 +447,26 @@ class TestGetWindowsDDNPMReleaseJsonInfo(unittest.TestCase):
 
 class TestGetReleaseJsonInfoForNextRC(unittest.TestCase):
     test_release_json = {
-        release.nightly_entry_for(6): {
+        nightly_entry_for(6): {
             "VERSION": "ver6_nightly",
             "HASH": "hash6_nightly",
         },
-        release.nightly_entry_for(7): {
+        nightly_entry_for(7): {
             "VERSION": "ver7_nightly",
             "HASH": "hash7_nightly",
         },
-        release.release_entry_for(6): {
+        release_entry_for(6): {
             "VERSION": "ver6_release",
             "HASH": "hash6_release",
         },
-        release.release_entry_for(7): {
+        release_entry_for(7): {
             "VERSION": "ver7_release",
             "HASH": "hash7_release",
         },
     }
 
     def test_get_release_json_info_for_next_rc_on_first_rc(self):
-        previous_release_json = release._get_release_json_info_for_next_rc(self.test_release_json, 7, True)
+        previous_release_json = _get_release_json_info_for_next_rc(self.test_release_json, 7, True)
 
         self.assertEqual(
             previous_release_json,
@@ -467,7 +477,7 @@ class TestGetReleaseJsonInfoForNextRC(unittest.TestCase):
         )
 
     def test_get_release_json_info_for_next_rc_on_second_rc(self):
-        previous_release_json = release._get_release_json_info_for_next_rc(self.test_release_json, 7, False)
+        previous_release_json = _get_release_json_info_for_next_rc(self.test_release_json, 7, False)
 
         self.assertEqual(
             previous_release_json,
@@ -480,26 +490,26 @@ class TestGetReleaseJsonInfoForNextRC(unittest.TestCase):
 
 class TestGetJMXFetchReleaseJsonInfo(unittest.TestCase):
     test_release_json = {
-        release.nightly_entry_for(6): {
+        nightly_entry_for(6): {
             "JMXFETCH_VERSION": "ver6_nightly",
             "JMXFETCH_HASH": "hash6_nightly",
         },
-        release.nightly_entry_for(7): {
+        nightly_entry_for(7): {
             "JMXFETCH_VERSION": "ver7_nightly",
             "JMXFETCH_HASH": "hash7_nightly",
         },
-        release.release_entry_for(6): {
+        release_entry_for(6): {
             "JMXFETCH_VERSION": "ver6_release",
             "JMXFETCH_HASH": "hash6_release",
         },
-        release.release_entry_for(7): {
+        release_entry_for(7): {
             "JMXFETCH_VERSION": "ver7_release",
             "JMXFETCH_HASH": "hash7_release",
         },
     }
 
     def test_get_release_json_info_for_next_rc_on_first_rc(self):
-        jmxfetch_version, jmxfetch_hash = release._get_jmxfetch_release_json_info(self.test_release_json, 7, True)
+        jmxfetch_version, jmxfetch_hash = _get_jmxfetch_release_json_info(self.test_release_json, 7, True)
 
         self.assertEqual(jmxfetch_version, "ver7_nightly")
         self.assertEqual(jmxfetch_hash, "hash7_nightly")
