+++ tasks/unit-tests/utils_tests.py
@@ -1,10 +1,7 @@
-import os
 import unittest
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock
 
-from invoke import MockContext, Result
-
-from tasks.libs.common.utils import clean_nested_paths, guess_from_keywords, guess_from_labels, query_version
+from tasks.libs.common.utils import clean_nested_paths, guess_from_keywords, guess_from_labels
 
 
 class TestUtils(unittest.TestCase):
@@ -68,60 +65,3 @@ class TestGuessFromKeywords(unittest.TestCase):
     def test_no_match(self):
         issue = MagicMock(title="fix bug", body="It comes from the file... hm I don't know.")
         self.assertEqual(guess_from_keywords(issue), "triage")
-
-
-class TestQueryVersion(unittest.TestCase):
-    @patch.dict(os.environ, {"BUCKET_BRANCH": "dev"}, clear=True)
-    def test_on_dev_bucket(self):
-        major_version = "7"
-        c = MockContext(
-            run={
-                r'git describe --tags --candidates=50 --match "7\.*" --abbrev=7': Result(
-                    "7.54.0-dbm-mongo-0.1-163-g315e3a2"
-                )
-            }
-        )
-        v, p, c, g, _ = query_version(c, major_version)
-        self.assertEqual(v, "7.54.0")
-        self.assertEqual(p, "dbm-mongo-0.1")
-        self.assertEqual(c, 163)
-        self.assertEqual(g, "315e3a2")
-
-    @patch.dict(os.environ, {"BUCKET_BRANCH": "nightly"}, clear=True)
-    def test_on_nightly_bucket(self):
-        major_version = "7"
-        c = MockContext(
-            run={
-                "git rev-parse --abbrev-ref HEAD": Result("main"),
-                rf"git tag --list --merged main | grep -E '^{major_version}\.[0-9]+\.[0-9]+(-rc.*|-devel.*)?$' | sort -rV | head -1": Result(
-                    "7.55.0-devel"
-                ),
-                'git describe --tags --candidates=50 --match "7.55.0-devel" --abbrev=7': Result(
-                    "7.55.0-devel-543-g315e3a2"
-                ),
-            }
-        )
-        v, p, c, g, _ = query_version(c, major_version)
-        self.assertEqual(v, "7.55.0")
-        self.assertEqual(p, "devel")
-        self.assertEqual(c, 543)
-        self.assertEqual(g, "315e3a2")
-
-    def test_on_release(self):
-        major_version = "7"
-        c = MockContext(
-            run={
-                "git rev-parse --abbrev-ref HEAD": Result("7.55.x"),
-                rf"git tag --list --merged 7.55.x | grep -E '^{major_version}\.[0-9]+\.[0-9]+(-rc.*|-devel.*)?$' | sort -rV | head -1": Result(
-                    "7.55.0-devel"
-                ),
-                'git describe --tags --candidates=50 --match "7.55.0-devel" --abbrev=7': Result(
-                    "7.55.0-devel-543-g315e3a2"
-                ),
-            }
-        )
-        v, p, c, g, _ = query_version(c, major_version, release=True)
-        self.assertEqual(v, "7.55.0")
-        self.assertEqual(p, "devel")
-        self.assertEqual(c, 543)
-        self.assertEqual(g, "315e3a2")
